refactor(#245): features/todo/ui 영역 리팩토링및 테스트 작성
