+++ .storybook/preview.tsx
@@ -1,4 +1,4 @@
-import "../src/app/globals.css";
+import "@/app/globals.css";
 
 import type { Preview } from "@storybook/nextjs-vite";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
