+++ src/shared/ui/Order/index.ts
@@ -1,0 +1,1 @@
+export { Order } from "./Order";
